+++ packages/svelte/src/index.js
@@ -1,4 +1,5 @@
 import Link from "./Link.html";
+import Navigating from "./Navigating.html";
 import curiStore from "./store.js";
 
-export { Link, curiStore };
+export { Link, Navigating, curiStore };